yall want more tests? me neither

+++ BACKEND/app/controllers/users_controller.rb
@@ -25,36 +25,49 @@ class UsersController < ApplicationController
     end
 
     def logoff
-        session_token = request.headers['Authorization']&.split(' ')&.last
-        user = User.joins(:tokens).where("tokens.token = '#{session_token}'").first
-        if params[:a]
-            tokens = Token.where(user_id: user.id)
-            tokens.each do |t|
-                t.destroy
+        begin
+            session_token = request.headers['Authorization']&.split(' ')&.last
+            user = User.joins(:tokens).where("tokens.token = '#{session_token}'").first
+            if !params[:a]
+                tokens = Token.where(user_id: user.id)
+                tokens.each do |t|
+                    t.destroy
+                end
+            else
+                Token.find_by(token: session_token).destroy
             end
-        else
-            Token.find_by(token: session_token).destroy
+        rescue NoMethodError
+            head 401
         end
     end
 
     def info
-        session_token = request.headers['Authorization']&.split(' ')&.last
-        user = User.joins(:tokens).where("tokens.token = '#{session_token}'").first
-        render json: {
-            :email => user.email,
-            :name => user.name,
-            :tag => user.tag,
-            :creation_date => user.created_at,
-            :theme => user.json["theme"],
-            :id => user.id
-        }
+        begin
+            session_token = request.headers['Authorization']&.split(' ')&.last
+            user = User.joins(:tokens).where("tokens.token = '#{session_token}'").first
+            render json: {
+                :email => user.email,
+                :name => user.name,
+                :tag => user.tag,
+                :creation_date => user.created_at,
+                :theme => user.json["theme"],
+                :id => user.id
+            }
+        rescue NoMethodError
+            head 401
+        end
+        
     end
 
     def update
+        session = request.headers['Authorization']
+        if session.nil?
+            head 401
+        end
         session_token = request.headers['Authorization']&.split(' ')&.last
         user = User.joins(:tokens).where("tokens.token = '#{session_token}'").first
         if session_token.nil? || user.nil?
-            render json: {:error => ERR_USER_NOT_EXIST}
+            render json: {:error => ERR_USER_NOT_EXIST}, :status => 404
         else
             user.update(name: params[:n]) unless params[:n].nil?
             if user.authenticate(params[:op])
@@ -72,18 +85,32 @@ class UsersController < ApplicationController
     end
 
     def destroy
+        session = request.headers['Authorization']
+        if session.nil?
+            head 401
+            return
+        end
         session_token = request.headers['Authorization']&.split(' ')&.last
         user = User.joins(:tokens).where("tokens.token = '#{session_token}'").first
         if user.nil?
-            render json: {:error => ERR_USER_NOT_EXIST}
+            render json: {:error => ERR_USER_NOT_EXIST}, :status => 401
         else
             user.destroy
         end
     end
 
     def projects
+        session = request.headers['Authorization']
+        if session.nil?
+            head 401
+            return
+        end
         session_token = request.headers['Authorization']&.split(' ')&.last
         user = User.joins(:tokens).where("tokens.token = '#{session_token}'").first
+        if user.nil?
+            head 401
+            return
+        end
         user_projects = ProjectsUser.joins(:project).where(user_id: user.id)
         if params[:search].present?
             user_projects = user_projects.where("projects.name LIKE ? OR projects.description LIKE ?", "%#{params[:search]}%", "%#{params[:search]}%")
